clean (FirmwareInstaller.cs): split up the native-proxy into its own partial
